gestion comment frontend

+++ Skill-Forge-Back/controllers/commentController.js
@@ -1,13 +1,16 @@
-// commentController.js
-
+const TrainingContent = require('../models/trainingContent');
 const Comment = require('../models/comment');
 
-// Créer un commentaire
+// Créer un commentaire et le pousser dans le contenu de formation
 exports.createComment = async (req, res) => {
   try {
-    const { content, userId, trainingContentId } = req.body;
-    const comment = new Comment({ content, user: userId, trainingContent: trainingContentId });
+    const { content,username, userId, trainingContentId } = req.body;
+    const comment = new Comment({ content, username, userId, trainingContent: trainingContentId });
     await comment.save();
+
+    // Pousser le commentaire nouvellement créé dans le contenu de formation associé
+    await TrainingContent.findByIdAndUpdate(trainingContentId, { $push: { comments: comment._id } });
+
     res.status(201).json(comment);
   } catch (err) {
     res.status(500).json({ error: err.message });
@@ -18,7 +21,7 @@ exports.createComment = async (req, res) => {
 exports.getCommentsByTrainingContentId = async (req, res) => {
   try {
     const { trainingContentId } = req.params;
-    const comments = await Comment.find({ trainingContent: trainingContentId }).populate();
+    const comments = await Comment.find({ trainingContent: trainingContentId }).populate('userId');
     res.json(comments);
   } catch (err) {
     res.status(500).json({ error: err.message });
@@ -37,10 +40,19 @@ exports.updateComment = async (req, res) => {
   }
 };
 
-// Supprimer un commentaire
+// Supprimer un commentaire et le retirer du contenu de formation associé
 exports.deleteComment = async (req, res) => {
   try {
     const { commentId } = req.params;
+    const comment = await Comment.findById(commentId);
+    if (!comment) {
+      return res.status(404).json({ message: 'Comment not found' });
+    }
+
+    // Retirer le commentaire de la liste des commentaires dans le contenu de formation associé
+    await TrainingContent.findByIdAndUpdate(comment.trainingContent, { $pull: { comments: commentId } });
+
+    // Supprimer le commentaire
     await Comment.findByIdAndDelete(commentId);
     res.json({ message: 'Comment deleted successfully' });
   } catch (err) {
